STEP 1: 파일 폴더링
