+++ ParcelInvoiceMaker/ParcelInvoiceMaker/Models/Repository/DatabaseParcelInformationPersistence.swift
@@ -1,0 +1,14 @@
+//
+//  DatabaseParcelInformationPersistence.swift
+//  ParcelInvoiceMaker
+//
+//  Created by 홍은표 on 3/7/24.
+//
+
+import Foundation
+
+final class DatabaseParcelInformationPersistence: ParcelInformationPersistence {
+  func save(parcelInformation: ParcelInformation) {
+    print("발송 정보를 데이터베이스에 저장했습니다.")
+  }
+}
