+++ ParcelInvoiceMaker/ParcelInvoiceMaker/Resources/Assets.xcassets/post_office_logo.imageset/Contents.json
