+++ ParcelInvoiceMaker/ParcelInvoiceMaker/Resources/Assets.xcassets/stamp.imageset/Contents.json
